Clean up warnings, mostly those that are errors that occur only when compiling for optimization -- all in service of re-enabling optimized builds
Makefile: respect build options for RCC platforms
os/interfaces/include/OcpiOsAssert.h: add two new conveniences:  ocpiAbort and ocpiIgnore
-- for dealing with nasty glibc insistent warnings etc.
os/interfaces/include/OcpiOsSocket.h: warnings, plus hooks for tcp buffer size and no-delay options
os/linux/include/OcpiOsPosixSocket.h: warnngs
os/linux/src/OcpiOsDebug.cxx: leave debug code commented out for delayed logging
os/linux/src/OcpiOsEther.cxx: warnings
os/linux/src/OcpiOsPosixSocket.cxx: warnngs
os/linux/src/OcpiOsSocket.cxx: warnings, set tcp buffer sizes, set tcp ndelay
projects/assets/components/util_comps/specs/test_source-spec.xml: add valuesSent volatile to see progress
projects/assets/components/util_comps/test_source.rcc/test_source.cc: ditto
runtime/application/src/OcpiApplication.cxx: warnings under optimization
runtime/application/src/ocpirun_main.cxx: ditto
runtime/dataplane/transport/include/TransportController.hh: ditto
runtime/dataplane/transport/include/TransportPortSet.hh: ditto
runtime/dataplane/transport/src/OcpiPort.cxx: ditto
runtime/dataplane/transport/src/TransportController.cc: ditto
runtime/dataplane/xfer/drivers/datagram/src/DtDataGramUDPDriver.cxx: warning
runtime/dataplane/xfer/drivers/socket/src/DtSocketXfer.cxx: resize receive buffer
runtime/dataplane/xfer/tests/src/transportTest_main.cxx:  warnings under optimization
runtime/drc/ad9361/src/RadioCtrlrNoOSTuneResamp.cc: ditto
runtime/drc/ad9361/src/ad9361_platform.cc: ditto
runtime/drc/ad9361/src/calc_ad9361_rf_tx_pll.cxx: fix bad return-the-address-on-the-stack warning
runtime/drc/base/src/RadioCtrlrConfigurator.cc: warnings
runtime/hdl-support/xilinx/xilinx.mk: fix broken warning when SDK cannot be found
runtime/hdl/include/HdlContainer.h: warnings
runtime/hdl/include/HdlWciControl.h: warnings
runtime/hdl/src/HdlBusDriver.cxx: remove debug code
runtime/hdl/src/HdlContainer.cxx: warnings
runtime/hdl/src/HdlSdp.cxx: warnings
runtime/library/src/OcpiLibraryAssembly.cxx: an actual crash bug when platform s misspelled
runtime/util/driver/src/OcpiDriverManager.cxx: warnings
runtime/util/misc/include/OcpiUtilMisc.h: warnings and broken macro when NDEBUG=1
tools/include/platform-defaults.mk: fix debug flags to be the right polarity
tools/ocpigen/src/hdl-worker.cxx: warnings
tools/ocpigen/src/wti.cxx: warnings
os/linux/src/OcpiOsFileSystem.cxx: warnings

+++ runtime/dataplane/xfer/drivers/datagram/src/DtDataGramUDPDriver.cxx
@@ -181,7 +181,7 @@ namespace DataTransfer {
 	m_msghdr.msg_name = &static_cast<EndPoint *>(&destEp)->sockaddr();
 	// We are depending on structure compatibility
 	m_msghdr.msg_iov = (struct iovec *)frame.iov;
-	m_msghdr.msg_iovlen = frame.iovlen;
+	m_msghdr.msg_iovlen = (int)frame.iovlen;
 	m_server.sendmsg(&m_msghdr, 0);
       }
       size_t
